+++ src/core/settings.cpp
@@ -19,10 +19,10 @@ void _setBrightness(uint8_t brightval) { }
 **  Function: setBrightness
 **  set brightness value
 **********************************************************************/
-void setBrightness(int brightval, bool save) {
+void setBrightness(uint8_t brightval, bool save) {
   if(bruceConfig.bright>100) bruceConfig.setBright(100);
-    _setBrightness(brightval);
-    delay(10);
+  _setBrightness(brightval);
+  delay(10);
 
   if(save){
     bruceConfig.setBright(brightval);
@@ -83,11 +83,11 @@ void setBrightnessMenu() {
   else if(bruceConfig.bright== 1) idx=4;
 
   options = {
-    {"100%", [=]() { setBrightness(100); }, bruceConfig.bright == 100 },
-    {"75 %", [=]() { setBrightness(75);  }, bruceConfig.bright == 75 },
-    {"50 %", [=]() { setBrightness(50);  }, bruceConfig.bright == 50 },
-    {"25 %", [=]() { setBrightness(25);  }, bruceConfig.bright == 25 },
-    {" 1 %", [=]() { setBrightness(1);   }, bruceConfig.bright == 1 },
+    {"100%", [=]() { setBrightness((uint8_t)100); }, bruceConfig.bright == 100 },
+    {"75 %", [=]() { setBrightness((uint8_t)75);  }, bruceConfig.bright == 75 },
+    {"50 %", [=]() { setBrightness((uint8_t)50);  }, bruceConfig.bright == 50 },
+    {"25 %", [=]() { setBrightness((uint8_t)25);  }, bruceConfig.bright == 25 },
+    {" 1 %", [=]() { setBrightness((uint8_t)1);   }, bruceConfig.bright == 1 },
     {"Main Menu", [=]() { backToMenu(); }}, // this one bugs the brightness selection
   };
   delay(200);
